refactor: comments 관련 API 함수 생성
- api 인스턴스에 patch 가 없어서 추가
- comments 관련 API 함수 추가
- API 함수에서 body에 보내는 파라미터값명이 고민됨.. request 식으로 해줘야할까? 아니면 그냥 형태로 두는게 나을까
